refactor(DecodePhone): 重构DecodePhone组件
