+++ src/decode-phone/__tests__/index.test.tsx
@@ -1,24 +1,31 @@
 import React from 'react';
-import { testMount, getCurrentRef, sleep } from '../../../tests';
-import { render, waitFor, act } from '@testing-library/react';
-// import { act } from 'react-dom/test-utils';
-import DecodePhone, { INIT_TITLE } from '..';
+import { act } from 'react-dom/test-utils';
 import { mount } from 'enzyme';
+import { testMount, getCurrentRef, sleep } from '../../../tests';
+import DecodePhone, { INIT_TITLE } from '..';
+import { decodePhone } from '../../shared/service';
+
+jest.mock('../../shared/service/index.ts', () => {
+  return {
+    decodePhone: jest.fn(),
+  };
+});
 
 const params = 'jZgOvsexafxJUlU3WHaMfA==';
-const testedEncodedTel = '155****1234';
+const testee = '155****1234';
 describe('🧪  DecodePhone', () => {
   testMount(DecodePhone);
-  // testSnapshot(DecodePhone);
 
   it('render content corrrectly', async (done) => {
+    decodePhone.mockResolvedValueOnce('ok');
     const onReadyHandler = jest.fn();
     const onErrorHandler = jest.fn();
     const wrapper = mount(
-      <DecodePhone params={params} onReady={onReadyHandler} onError={onErrorHandler}>
-        <span id="phone">{testedEncodedTel}</span>
+      <DecodePhone params={params} onReady={onReadyHandler}>
+        <span id="phone">{testee}</span>
       </DecodePhone>,
     );
+
     wrapper
       .find('#phone')
       .at(0)
@@ -27,22 +34,5 @@ describe('🧪  DecodePhone', () => {
     expect(onReadyHandler).toHaveBeenCalledTimes(1);
     expect(onErrorHandler).toHaveBeenCalledTimes(0);
     expect(getCurrentRef(wrapper, '.ant-tooltip-inner').innerHTML).not.toBe(INIT_TITLE);
-  });
-
-  it('second click will not do request', async () => {
-    const onReadyHandler = jest.fn();
-    const onErrorHandler = jest.fn();
-    const wrapper = mount(
-      <DecodePhone params={params} onReady={onReadyHandler} onError={onErrorHandler}>
-        <span id="phone">{testedEncodedTel}</span>
-      </DecodePhone>,
-    );
-    const div = wrapper.find('#phone').at(0);
-    div.simulate('click');
-    expect(onReadyHandler).toHaveBeenCalledTimes(1);
-    expect(onErrorHandler).toHaveBeenCalledTimes(1);
-    div.simulate('click');
-    div.simulate('click');
-    expect(onReadyHandler).toHaveBeenCalledTimes(1);
   });
 });
