+++ src/decode-phone/index.tsx
@@ -1,31 +1,35 @@
-import React, { useState, useEffect } from 'react';
-import { TooltipProps } from 'antd/lib/tooltip';
-import { Tooltip } from 'antd';
-import { Tel, DECODE_URL } from '../shared/service/decode-phone/model';
+import React, { useState } from 'react';
+import { TooltipProps, RenderFunction } from 'antd/lib/tooltip';
+import { Tooltip, message } from 'antd';
+import { decodePhone } from '../shared/service';
 
-export const ERROR_TITLE = '手机号解码失败';
 export const INIT_TITLE = '加载中';
 interface DecodePhoneProps extends Omit<TooltipProps, 'title'> {
-  url?: string;
+  // 待解手机号码
   params: string;
+  // 解码成功时的回调
   onReady?(tel): void;
+  // 解码错误时的回调
   onError?(err): void;
+  // 更改解码手机号的请求地址
+  url?: string;
 }
 
 const DecodePhone = (props: DecodePhoneProps) => {
-  const { children, params, url = DECODE_URL, onReady, onError, ...rest } = props;
-  const [title, setTitle] = useState(INIT_TITLE as any);
-
-  const decode = () => {
-    const phone = new Tel(params);
-    phone.decodePhone().then(() => {
-      const tel = phone.toString();
-      tel && setTitle(tel);
-    });
+  const { children, params, url, onReady, onError, ...rest } = props;
+  const [title, setTitle] = useState<RenderFunction | React.ReactNode>(INIT_TITLE);
+  const handleRequest = async () => {
+    try {
+      const tel = await decodePhone(params, url);
+      setTitle(tel);
+      onReady?.(tel);
+    } catch (err) {
+      onError?.(err);
+      message.error('network error', err);
+    }
   };
-
   return (
-    <Tooltip {...rest} title={title} trigger="click" onVisibleChange={(visible) => visible && [INIT_TITLE, ERROR_TITLE].includes(title) && decode()}>
+    <Tooltip {...rest} title={title} trigger="click" onVisibleChange={(visible) => visible && title === INIT_TITLE && handleRequest()}>
       {children}
     </Tooltip>
   );
