+++ src/shared/service/constants.ts
@@ -1,8 +1,8 @@
 export type supporttingTypes = 'production' | 'development' | 'test' | 'staging' | 'dev';
-enum env {
+
+export enum env {
   'production' = '',
   'development' = 'dev',
   'test' = 'test',
   'staging' = 'staging',
 }
-export const envTransformer = (NODE_ENV: supporttingTypes) => env[NODE_ENV];
