+++ tests/sleep.ts
@@ -1,8 +1,6 @@
 import { act } from 'react-dom/test-utils';
-
-const globalTimeout = global.setTimeout;
 export default async (timeout = 0) => {
   await act(async () => {
-    await new Promise((resolve) => globalTimeout(resolve, timeout));
+    await new Promise((resolve) => setTimeout(resolve, timeout));
   });
 };
